+++ shared/src/index.ts
@@ -3,6 +3,7 @@ export * from "./league";
 export * from "./valuation";
 export * from "./player";
 export * from "./goalkeeperGrid";
+export * from "./probableLineup";
 export * from "./manager";
 export * from "./purchase";
 export * from "./maxBid";